Fix minor issues

+++ Shopping/Shared/Requests/Statistic/GetExpensesByKind.cs
@@ -2,7 +2,7 @@
 
 namespace Shopping.Shared.Requests.Statistic
 {
-    public sealed class GetExpensesByKind : IRequest<IDictionary<string, double>>
+    public sealed class GetExpensesByKind : IRequest<IDictionary<string, decimal>>
     {
         public DateTime StartOfMonth { get; set; }
     }
