+++ Shopping/Shared/Requests/Statistic/GetExpensesByMonth.cs
@@ -2,7 +2,7 @@
 
 namespace Shopping.Shared.Requests.Statistic
 {
-    public sealed class GetExpensesByMonth : IRequest<IDictionary<int, double>>
+    public sealed class GetExpensesByMonth : IRequest<IDictionary<int, decimal>>
     {
         public DateTime StartOfYear { get; set; }
     }
