+++ Shopping/Shared/Requests/Statistic/GetExpensesByShop.cs
@@ -2,7 +2,7 @@
 
 namespace Shopping.Shared.Requests.Statistic
 {
-    public sealed class GetExpensesByShop : IRequest<IDictionary<string, double>>
+    public sealed class GetExpensesByShop : IRequest<IDictionary<string, decimal>>
     {
         public DateTime StartOfMonth { get; set; }
     }
