+++ Shopping/Shopping/Services/Handlers/Statistic/GetExpensesByMonthHandler.cs
@@ -5,7 +5,7 @@ using Shopping.Shared.Requests.Statistic;
 
 namespace Shopping.Services.Handlers.Statistic
 {
-    public sealed class GetExpensesByMonthHandler : IRequestHandler<GetExpensesByMonth, IDictionary<int, double>>
+    public sealed class GetExpensesByMonthHandler : IRequestHandler<GetExpensesByMonth, IDictionary<int, decimal>>
     {
         private readonly ShoppingDbContext _context;
 
@@ -14,7 +14,7 @@ namespace Shopping.Services.Handlers.Statistic
             _context = context;
         }
 
-        public async Task<IDictionary<int, double>> Handle(GetExpensesByMonth request, CancellationToken cancellationToken)
+        public async Task<IDictionary<int, decimal>> Handle(GetExpensesByMonth request, CancellationToken cancellationToken)
         {
             var firstDayOfYear = request.StartOfYear;
             var lastDayOfYear = request.StartOfYear.AddYears(1).AddSeconds(-1);
@@ -25,7 +25,7 @@ namespace Shopping.Services.Handlers.Statistic
                         select new { Month = receipt.Date.Month, Price = receiptItem.Price, Amount = receiptItem.Amount };
 
             var data = await query.OrderBy(x => x.Month).ToArrayAsync(cancellationToken).ConfigureAwait(false);
-            return data.GroupBy(d => d.Month).ToDictionary(x => x.Key, v => v.Sum(d => (double)(d.Price * d.Amount)));
+            return data.GroupBy(d => d.Month).ToDictionary(x => x.Key, v => v.Sum(d => (d.Price * d.Amount)));
         }
     }
 }
